requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2
django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
pillow==9.3.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.2.1
setuptools==65.6.3; python_version >= '3.7'
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
